requirements.txt
pyvcf==0.6.*
